checkpoint: renames "Attempt/error" to "Attempt/Error"
