+++ src/library/Attempt/Adapted/Config.ts
@@ -1,8 +1,8 @@
 import type {
   Attempt_Error_Actionable,
-} from '../error';
+} from '../Error';
 
-import type Attempt_Error_Interpreter from '../error/Interpreter';
+import type Attempt_Error_Interpreter from '../Error/Interpreter';
 
 interface Attempt_Adapted_Config<
   SomeActionableError extends Attempt_Error_Actionable<string>,
