+++ src/library/Attempt/Error/NonActionableBuiltInError.ts
