+++ src/library/Attempt/Error/NonActionableError.ts
