+++ src/library/Attempt/Error/assertions/assertAppropriatelyThrown.ts
