+++ src/library/Attempt/Error/assertions/index.ts
