+++ src/library/Attempt/Error/index.ts
