+++ src/library/Attempt/Error/modifier/AppropriatelyThrown.ts
