+++ src/library/Attempt/Outcome/exports.ts
@@ -1,6 +1,6 @@
 import type {
   Attempt_Error_Actionable,
-} from '../error';
+} from '../Error';
 
 import {
   type Attempt_Outcome_Failure,
