+++ src/library/Attempt/factory/AttemptCreationError.ts
@@ -1,4 +1,4 @@
-import Attempt_Error_NonActionable from '../error/NonActionableError';
+import Attempt_Error_NonActionable from '../Error/NonActionableError';
 
 class Attempt_Error_Creation
   extends Attempt_Error_NonActionable {
